ClusteredNeuroVol

Three-dimensional brain image that is divided into N disjoint partitions

Construct a ClusteredNeuroVol instance

ClusteredNeuroVol(mask, clusters, label_map = NULL, label = "")

Arguments

mask
an instance of class LogicalNeuroVol

clusters
a vector of clusters ids with length equal to number of nonzero voxels in mask mask

label_map
an optional list that maps from cluster id to a cluster label, e.g. (1 -> "FFA", 2 -> "PPA")

label
an optional character string used to label of the volume

Value

ClusteredNeuroVol instance

Slots

mask

the logical mask indicating the spatial domain of the set of clusters

clusters

an integer index indicating the cluster number for each voxel in the mask

label_map

a list mapping from name to cluster id

cluster_map

an environment mapping from cluster id to the set of 1D spatial indices

Examples

bspace <- NeuroSpace(c(16,16,16), spacing=c(1,1,1))
grid <- index_to_grid(bspace, 1:(16*16*16))
kres <- kmeans(grid, centers=10)
mask <- NeuroVol(rep(1, 16^3),bspace)
clusvol <- ClusteredNeuroVol(mask, kres$cluster)