bootstrap_searchlight

Create a spherical searchlight iterator that samples regions from within a mask.

Create an exhaustive searchlight iterator that only returns voxel coordinates.

Generates an ROIVol around each non-zero voxel in a 3D image mask

bootstrap_searchlight(mask, radius = 8, iter = 100)

searchlight_coords(mask, radius, nonzero = FALSE, cores = 0)

searchlight(mask, radius, eager = FALSE, nonzero = FALSE, cores = 0)

clustered_searchlight(mask, cvol = NULL, csize = NULL)

Arguments

mask

an image volume containing valid central voxels for roving searchlight.

radius

in mm of spherical searchlight.

iter

the total number of searchlights to sample (default is 100).

nonzero

only include coordinates with nonzero values in the supplied mask.

cores

number of cores to use

eager

if TRUE, then all searchlight coordinates set are generated up front. This is faster but requires more memory to store all coordinates.

cvol

a ClusteredNeuroVol instance

csize

the number of clusters (ignored if cvol is provided)

Value

a list of matrices containing of integer-valued voxel coordinates. a lazy list of ROIVolWindow objects an iter class

Details

searchlight centers are sampled without replacement, but the same surround voxel can belong to multiple searchlight samples.

Examples

mask <- read_vol(system.file("extdata", "global_mask.nii", package="neuroim2"))
slight <- bootstrap_searchlight(mask, 6)