Arithmetic Operations

Methods for performing arithmetic operations on neuroimaging objects

This method performs arithmetic operations between two ROIVol objects (e1 and e2) using a generic arithmetic function. The dimensions of both objects are checked for compatibility before performing the operation.

Perform an arithmetic operation between two SparseNeuroVec objects. The input SparseNeuroVec objects must have the same dimensions and NeuroSpace objects. The method computes the union of the masks and performs the arithmetic operation on the non-zero values. The result is returned as a new SparseNeuroVec object.

Perform an arithmetic operation between a SparseNeuroVol object and a NeuroVol object. The input SparseNeuroVol and NeuroVol objects must have the same dimensions. The method performs the arithmetic operation on the non-zero values of the SparseNeuroVol and the corresponding values of the NeuroVol. The result is returned as a new DenseNeuroVol object.

Perform an arithmetic operation between a NeuroVol object and a SparseNeuroVol object. The input NeuroVol and SparseNeuroVol objects must have the same dimensions. The method performs the arithmetic operation on the values of the NeuroVol and the non-zero values of the SparseNeuroVol. The result is returned as a new DenseNeuroVol object.

Perform an arithmetic operation between two NeuroVec objects. The input NeuroVec objects must have the same dimensions. The method performs the arithmetic operation on the elements of the NeuroVec objects. The result is returned as a new DenseNeuroVec object.

Perform an arithmetic operation between a NeuroVol object and a SparseNeuroVol object. The input NeuroVol and SparseNeuroVol objects must have the same dimensions. The method performs the arithmetic operation on the values of the NeuroVol and the non-zero values of the SparseNeuroVol. The result is returned as a new DenseNeuroVol object.

Usage

# S4 method for class 'SparseNeuroVol,SparseNeuroVol'
Arith(e1, e2)

# S4 method for class 'DenseNeuroVol,DenseNeuroVol'
Arith(e1, e2)

# S4 method for class 'DenseNeuroVec,DenseNeuroVec'
Arith(e1, e2)

# S4 method for class 'SparseNeuroVol,NeuroVol'
Arith(e1, e2)

# S4 method for class 'NeuroVol,SparseNeuroVol'
Arith(e1, e2)

# S4 method for class 'NeuroVec,NeuroVec'
Arith(e1, e2)

# S4 method for class 'NeuroVol,SparseNeuroVol'
Arith(e1, e2)

# S4 method for class 'SparseNeuroVec,SparseNeuroVec'
Arith(e1, e2)

Arguments

e1

A NeuroVol object.

e2

A SparseNeuroVol object.

Value

A DenseNeuroVol object representing the result of the arithmetic operation.

An ROIVol object containing the result of the arithmetic operation between e1 and e2.

A SparseNeuroVec object representing the result of the arithmetic operation.

A DenseNeuroVol object representing the result of the arithmetic operation.

A DenseNeuroVol object representing the result of the arithmetic operation.

A DenseNeuroVec object representing the result of the arithmetic operation.

A DenseNeuroVol object representing the result of the arithmetic operation.