Generic function to center/scale row-subsets of a matrix or matrix-like object

Generic function to center/scale row-subsets of a matrix or matrix-like object

split_scale(x, f, center, scale)

# S4 method for matrix,factor,logical,logical
split_scale(x, f,
  center = TRUE, scale = TRUE)

# S4 method for matrix,factor,missing,missing
split_scale(x, f)

# S4 method for DenseNeuroVec,factor,missing,missing
split_scale(x, f)

# S4 method for DenseNeuroVec,factor,logical,missing
split_scale(x, f, center)

# S4 method for DenseNeuroVec,factor,logical,logical
split_scale(x, f, center,
  scale)

Arguments

x
a numeric matrix or matrix-like object

f
the splitting object, typically a factor or set of integer indices. must be equal to number of rows of matrix.

center
should values within each submatrix be centered? (mean removed from each column of submatrix)

scale
should values be scaled? (divide vector by standard deviation from each column of submatrix)

Value

a new matrix or matrix-like object where the original rows have been grouped by f and then centered and/or scaled for each grouping

Examples

M <- matrix(rnorm(1000), 10, 100)
fac <- factor(rep(1:2, each=5))
Ms <- split_scale(M, fac)

## correctly centered
all(abs(apply(Ms[fac == 1,], 2, mean)) < .000001)

#> [1] TRUE

all(abs(apply(Ms[fac == 2,], 2, mean)) < .000001)

#> [1] TRUE

# correctly scaled
all.equal(apply(Ms[fac == 1,], 2, sd), rep(1, ncol(Ms)))

#> [1] TRUE

all.equal(apply(Ms[fac == 2,], 2, sd), rep(1, ncol(Ms)))

#> [1] TRUE